backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "fastapi[standard]>=0.117.1",
    "loguru>=0.7.3",
    "lxml>=6.0.1",
    "newsapi-python>=0.2.7",
    "pydantic>=2.11.9",
    "pymongo>=4.15.1",
    "readability-lxml>=0.8.4.1",
]
